[refactor] Se refactoriza HostProfile en varios componentes y hooks
